Modified API Methods

+++ src/main/java/omce/ws/OMCEWebController.java
@@ -10,6 +10,7 @@ import omce.ws.repositories.StudentGradeRepository;
 import omce.ws.repositories.StudentRepository;
 import omce.ws.utils.StudentGradeInfo;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -39,24 +40,30 @@ public class OMCEWebController {
     // POST
 
     @PostMapping("/createExam")
-    public ResponseEntity<?> createExam(@RequestBody Exam exam) {
+    public ResponseEntity<Exam> createExam(@RequestBody Exam exam) {
 
         Exam examDB = examRepository.findByContent(exam.getDescription(),
                 exam.getDate(), exam.getTime(), exam.getLocation());
         if (examDB != null)
-            return ResponseEntity.status(HttpStatus.OK).body(examDB);
-        Exam e = examRepository.save(exam);
-        return ResponseEntity.status(HttpStatus.CREATED).body(e);
+            return ResponseEntity.status(HttpStatus.CONFLICT).body(examDB);
+
+        try{
+            Exam e = examRepository.save(exam);
+            return ResponseEntity.status(HttpStatus.CREATED).body(e);
+        } catch (DataIntegrityViolationException e){
+            return ResponseEntity.status(HttpStatus.CONFLICT).body(null);
+        }
+
     }
 
     // Modify exam’s description.
     // PUT
-    @PutMapping("/updateExam/{examId}")
-    public ResponseEntity updateExam(@PathVariable Long examId, @RequestBody Exam exam) {
+    @PutMapping("/updateExamDescription/{examId}")
+    public ResponseEntity updateExamDescription(@PathVariable Long examId, @RequestBody String description) {
         Optional<Exam> examToEdit = examRepository.findById(examId);
         if (!examToEdit.isPresent())
-            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam is not present");
-        examToEdit.get().setDescription(exam.getDescription());
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam does not exist");
+        examToEdit.get().setDescription(description);
         examRepository.save(examToEdit.get());
         return ResponseEntity.status(HttpStatus.OK).body("Exam updated correctly");
     }
@@ -64,33 +71,38 @@ public class OMCEWebController {
     // Delete exam (if it has no grades).
     // DELETE
     @DeleteMapping("/deleteExam/{examId}")
-    public ResponseEntity deleteExam(@PathVariable Long examId, @RequestBody Exam exam) {
+    public ResponseEntity deleteExam(@PathVariable Long examId) {
         Optional<Exam> examToDelete = examRepository.findById(examId);
         if (!examToDelete.isPresent())
-            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam is not present");
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam does not exist");
 
-        examRepository.delete(examToDelete.get());
-        return ResponseEntity.status(HttpStatus.OK).body("Exam deleted correctly");
+        Optional<List<StudentGrade>> studentGrades = studentGradeRepository.findByExamId(examId);
+        if (!studentGrades.isPresent()){
+            examRepository.delete(examToDelete.get());
+            return ResponseEntity.status(HttpStatus.OK).body("Exam deleted correctly");
+        }
+
+        return ResponseEntity.status(HttpStatus.CONFLICT).body("Exam has grades");
     }
 
 
     // Search exam content using the key or search them using textual description (full/partial search).
     // GET
     @GetMapping(value = "/searchExamById/{id}")
-    public ResponseEntity<Exam> searchExamById(@PathVariable Long id) {
+    public ResponseEntity searchExamById(@PathVariable Long id) {
         Optional<Exam> exam = examRepository.findById(id);
         if (!exam.isPresent())
-            return ResponseEntity.noContent().build();
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam does not exist");
 
         return ResponseEntity.ok(exam.get());
     }
 
     @GetMapping("/searchExamByDescription/{description}")
-    public ResponseEntity<?> searchExamByDescription(@PathVariable String description) {
+    public ResponseEntity searchExamByDescription(@PathVariable String description) {
         List<Exam> exams = examRepository.findByDescription(description);
 
         if (exams.size() == 0)
-            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam is not present");
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam does not exist");
 
         return ResponseEntity.ok(exams);
     }
@@ -100,46 +112,44 @@ public class OMCEWebController {
     // GET
 
     @GetMapping("/downloadExam/{examId}")
-    public ResponseEntity<?> downloadExamById(@PathVariable Long examId) {
+    public ResponseEntity downloadExamById(@PathVariable Long examId) {
 
         Optional<Exam> exam = examRepository.findById(examId);
         if (!exam.isPresent())
-            return ResponseEntity.status(HttpStatus.CONFLICT).body("Error downloading the exam");
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam does not exist");
 
         try{
             ObjectMapper objectMapper = new ObjectMapper();
             String json = objectMapper.writeValueAsString(exam.get());
             byte[] examBytes = json.getBytes();
-            String fileName = "exam" + examId + ".json";
+            String fileName = "Exam" + examId + ".json";
             return ResponseEntity.ok()
                     .header(HttpHeaders.CONTENT_DISPOSITION, "attachment;filename=" + fileName)
                     .contentType(MediaType.APPLICATION_JSON)
                     .contentLength(examBytes.length).body(examBytes);
         } catch (JsonProcessingException e) {
-            e.printStackTrace();
             return ResponseEntity.status(HttpStatus.CONFLICT).body("Error downloading the exam");
         }
 
     }
 
     @GetMapping("/downloadExams")
-    public ResponseEntity<?> downloadExams() {
+    public ResponseEntity downloadExams() {
 
         List<Exam> exams = examRepository.findAll();
         if (exams.size() == 0)
-            return ResponseEntity.status(HttpStatus.CONFLICT).body("no hi ha exams");
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam does not exist");
 
         try{
             ObjectMapper objectMapper = new ObjectMapper();
             String json = objectMapper.writeValueAsString(exams);
             byte[] examBytes = json.getBytes();
-            String fileName = "exams.json";
+            String fileName = "Exams.json";
             return ResponseEntity.ok()
                     .header(HttpHeaders.CONTENT_DISPOSITION, "attachment;filename=" + fileName)
                     .contentType(MediaType.APPLICATION_JSON)
                     .contentLength(examBytes.length).body(examBytes);
         } catch (JsonProcessingException e) {
-            e.printStackTrace();
             return ResponseEntity.status(HttpStatus.CONFLICT).body("Error downloading the exams");
         }
     }
@@ -152,7 +162,7 @@ public class OMCEWebController {
     public ResponseEntity uploadGrades(@RequestBody List<StudentGradeInfo> studentGrades) {
 
         if (studentGrades.size() == 0)
-            return ResponseEntity.status(HttpStatus.CONFLICT).build();
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Request body is empty");
 
         try{
             for (StudentGradeInfo sg: studentGrades) {
@@ -163,19 +173,19 @@ public class OMCEWebController {
                 studentGradeRepository.save(studentGrade);
             }
         }catch (Exception e){
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Exam or student does not exist");
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Exam does not exist");
         }
-        return ResponseEntity.ok(HttpStatus.CREATED);
+        return ResponseEntity.status(HttpStatus.CREATED).body("Student grade created");
     }
 
     // Download student’s grades.
     // GET
-    @GetMapping("/downloadStudentGrades/{studentId}")
-    public ResponseEntity<?> downloadStudentGrades(@PathVariable String universityId) {
+    @GetMapping("/downloadStudentGrades/{universityId}")
+    public ResponseEntity downloadStudentGrades(@PathVariable String universityId) {
 
         List<StudentGrade> studentGrades = studentGradeRepository.findByUniversityId(universityId);
         if (studentGrades.size() == 0)
-            return ResponseEntity.status(HttpStatus.CONFLICT).body("no hi ha StudentGrades");
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Student does not exist");
 
         ArrayList<StudentGradeInfo> studentGradesInfo = new ArrayList<>();
         for (StudentGrade sg : studentGrades) {
@@ -188,13 +198,12 @@ public class OMCEWebController {
             ObjectMapper objectMapper = new ObjectMapper();
             String json = objectMapper.writeValueAsString(studentGradesInfo);
             byte[] examBytes = json.getBytes();
-            String fileName = "Grades" + universityId + ".json";
+            String fileName = universityId + "grades.json";
             return ResponseEntity.ok()
                     .header(HttpHeaders.CONTENT_DISPOSITION, "attachment;filename=" + fileName)
                     .contentType(MediaType.APPLICATION_JSON)
                     .contentLength(examBytes.length).body(examBytes);
         } catch (JsonProcessingException e) {
-            e.printStackTrace();
             return ResponseEntity.status(HttpStatus.CONFLICT).body("Error downloading grades");
         }
 
@@ -202,14 +211,19 @@ public class OMCEWebController {
 
     // Manage student’s access (by ID).
     // POST
-    @PostMapping("/createStudent/{universityId}")
-    public ResponseEntity createStudent(@PathVariable String universityId) {
+    @PostMapping("/createStudent")
+    public ResponseEntity createStudent(@RequestBody Student student) {
 
-        Student student = studentRepository.findByUniversityId(universityId);
-        if (student != null)
-            return ResponseEntity.status(HttpStatus.CONFLICT).build();
-        studentRepository.save(new Student(universityId));
-        return ResponseEntity.status(HttpStatus.CREATED).body("Student created");
+        Student studentDB = studentRepository.findByUniversityId(student.getUniversityId());
+        if (studentDB != null)
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Student is already created");
+
+        try{
+            studentRepository.save(new Student(student.getUniversityId()));
+            return ResponseEntity.status(HttpStatus.CREATED).body("Student created");
+        } catch (DataIntegrityViolationException e){
+            return ResponseEntity.status(HttpStatus.CONFLICT).body("Incorrect request body");
+        }
     }
 
     /* Integration Functions */
@@ -227,7 +241,7 @@ public class OMCEWebController {
                 exam.getDate(), exam.getTime(), exam.getLocation());
 
         if (examDB == null)
-            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam is not present");
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Exam does not exist");
 
         return ResponseEntity.ok(examDB);
     }
@@ -236,7 +250,7 @@ public class OMCEWebController {
     public ResponseEntity deleteStudent(@PathVariable String universityId) {
         Student studentToDelete = studentRepository.findByUniversityId(universityId);
         if (studentToDelete == null)
-            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Student is not present");
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).body("Student does not exist");
         studentRepository.delete(studentToDelete);
         return ResponseEntity.status(HttpStatus.OK).body("Student deleted correctly");
     }
